refactor: simplify how to choose the sql template
