+++ src/main/java/io/seqera/migtool/sql/PostgresSqlTemplate.java
@@ -1,38 +1,38 @@
-package io.seqera.migtool.strategy;
+package io.seqera.migtool.sql;
 
-public class NotPostgresQueryBuilder extends DialectQueryBuilder {
+public class PostgresSqlTemplate extends DefaultSqlTemplate {
 
-    public NotPostgresQueryBuilder(String tableName) {
+    public PostgresSqlTemplate(String tableName) {
         super(tableName);
     }
 
     @Override
     protected String id() {
-        return "`id`";
+        return "id";
     }
 
     @Override
     protected String rank() {
-        return "`rank`";
+        return "rank";
     }
 
     @Override
     protected String script() {
-        return "`script`";
+        return "script";
     }
 
     @Override
     protected String checksum() {
-        return "`checksum`";
+        return "checksum";
     }
 
     @Override
     protected String createdOn() {
-        return "`created_on`";
+        return "created_on";
     }
 
     @Override
     protected String executionTime() {
-        return "`execution_time`";
+        return "execution_time";
     }
 }
